The L2 stability is a joke

+++ src/dg/flux.cpp
@@ -57,7 +57,7 @@ std::array<double, 2> Flux::barenBlattFlux(double u, double q, double m)
     // if(m==1.0){
     //     return {(-1.0*q),(-1.0)};
     // }else{
-    return {(-sqrt(m*pow(u,m-1.0))*q),(-sqrt(double(m*pow(u,m-1.0))))};
+    return {(-std::sqrt(m*std::pow(u,m-1.0))*q),(-std::sqrt(m*std::pow(u,m-1.0)))};
     // }
 
 }
